+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [5.2.0] - 2024-01-18
+
 ### Added
 
 - `apiNSWLiquorLicense` form element to `typeCastService` and `submissionService`